refactor: extract PublisherCard for LN data
